Fix failing GoogleTestVerification unit test
I do not know the origin of this error---I have never encountered it before and
suppose that it is somehow related to using Cyclus v1.6, or to updates in the
gtest software.

Below is the full output of this unit test:

[----------] 1 test from GoogleTestVerification
[ RUN      ] GoogleTestVerification.UninstantiatedParameterizedTestSuite<InstitutionTests>
/home/ms290211/NVD/cyclus/tests/agent_tests/institution_tests.cc:6: Failure
Parameterized test suite InstitutionTests is defined via TEST_P, but never instantiated. None of the test cases will run. Either no INSTANTIATE_TEST_SUITE_P is provided or the only ones provided expand to nothing.

Ideally, TEST_P definitions should only ever be included as part of binaries that intend to use them. (As opposed to, for example, being placed in a library that may be linked in to get other utilities.)

To suppress this error for this test suite, insert the following line (in a non-header) in the namespace it is defined in:

GTEST_ALLOW_UNINSTANTIATED_PARAMETERIZED_TEST(InstitutionTests);

[  FAILED  ] GoogleTestVerification.UninstantiatedParameterizedTestSuite<InstitutionTests> (0 ms)
[----------] 1 test from GoogleTestVerification (0 ms total)